feat: add amplify status api -acm Model
This commit adds a command for viewing the ACM for @auth v2.
